init lab 6
