+++ bookmark_strategy.py
@@ -1,0 +1,8 @@
+class BookmarkStrategy:
+    def process_bookmarks(self, bookmarks):
+        pass
+
+
+class DefaultBookmarkStrategy(BookmarkStrategy):
+    def process_bookmarks(self, bookmarks):
+        pass
