+++ macro_strategy.py
@@ -1,0 +1,8 @@
+class MacroStrategy:
+    def execute_macros(self, macros):
+        pass
+
+
+class DefaultMacroStrategy(MacroStrategy):
+    def execute_macros(self, macros):
+        pass
