+++ snippet_strategy.py
@@ -1,0 +1,8 @@
+class SnippetStrategy:
+    def apply_snippets(self, snippets):
+        pass
+
+
+class DefaultSnippetStrategy(SnippetStrategy):
+    def apply_snippets(self, snippets):
+        pass
