+++ syntax_highlighter.py
@@ -1,0 +1,8 @@
+class SyntaxHighlighterStrategy:
+    def highlight(self, text):
+        pass
+
+
+class PythonSyntaxHighlighter(SyntaxHighlighterStrategy):
+    def highlight(self, text):
+        pass
